# [NEON] Faster UTF-8 conversion routines