--- a/src/arm64/arm_convert_utf8_to_utf16.cpp
+++ b/src/arm64/arm_convert_utf8_to_utf16.cpp
@@ -10,11 +10,6 @@ size_t convert_masked_utf8_to_utf16(const char *input,
   // Why 12 input bytes and not 16? Because we are concerned with the size of
   // the lookup tables. Also 12 is nicely divisible by two and three.
   //
-  #ifdef SIMDUTF_REGULAR_VISUAL_STUDIO
-  const uint8x16_t swap = make_uint8x16_t(1, 0, 3, 2, 5, 4, 7, 6, 9, 8, 11, 10, 13, 12, 15, 14);
-  #else
-  const uint8x16_t swap = {1, 0, 3, 2, 5, 4, 7, 6, 9, 8, 11, 10, 13, 12, 15, 14};
-  #endif
   uint8x16_t in = vld1q_u8(reinterpret_cast<const uint8_t*>(input));
   const uint16_t input_utf8_end_of_code_point_mask =
       utf8_end_of_code_point_mask & 0xfff;
@@ -22,15 +17,15 @@ size_t convert_masked_utf8_to_utf16(const char *input,
   // Optimization note: our main path below is load-latency dependent. Thus it is maybe
   // beneficial to have fast paths that depend on branch prediction but have less latency.
   // This results in more instructions but, potentially, also higher speeds.
-  //
+
   // We first try a few fast paths.
   if((utf8_end_of_code_point_mask & 0xffff) == 0xffff) {
     // We process in chunks of 16 bytes
-    uint16x8_t ascii_first = vmovl_u8(vget_low_u8 (in));
+    uint16x8_t ascii_first = vmovl_u8(vget_low_u8(in));
     uint16x8_t ascii_second = vmovl_high_u8(in);
     if (!match_system(big_endian)) {
-      ascii_first = vreinterpretq_u16_u8(vqtbl1q_u8(vreinterpretq_u8_u16(ascii_first), swap));
-      ascii_second = vreinterpretq_u16_u8(vqtbl1q_u8(vreinterpretq_u8_u16(ascii_second), swap));
+      ascii_first = vreinterpretq_u16_u8(vrev16q_u8(vreinterpretq_u8_u16(ascii_first)));
+      ascii_second = vreinterpretq_u16_u8(vrev16q_u8(vreinterpretq_u8_u16(ascii_second)));
     }
     vst1q_u16(reinterpret_cast<uint16_t*>(utf16_output), ascii_first);
     vst1q_u16(reinterpret_cast<uint16_t*>(utf16_output) + 8, ascii_second);
@@ -39,40 +34,96 @@ size_t convert_masked_utf8_to_utf16(const char *input,
   }
   if((utf8_end_of_code_point_mask & 0xffff) == 0xaaaa) {
     // We want to take 8 2-byte UTF-8 words and turn them into 8 2-byte UTF-16 words.
-    // There is probably a more efficient sequence, but the following might do.
-    uint8x16_t perm = vqtbl1q_u8(in, swap);
-    uint8x16_t ascii = vandq_u8(perm, vreinterpretq_u8_u16(vmovq_n_u16(0x7f)));
-    uint8x16_t highbyte = vandq_u8(perm, vreinterpretq_u8_u16(vmovq_n_u16(0x1f00)));
-    uint8x16_t composed = vorrq_u8(ascii, vreinterpretq_u8_u16(vshrq_n_u16(vreinterpretq_u16_u8(highbyte), 2)));
-    if (!match_system(big_endian)) composed = vqtbl1q_u8(composed, swap);
-    vst1q_u8(reinterpret_cast<uint8_t*>(utf16_output), composed);
+    // 10bbbbbb110aaaaa
+    uint16x8_t upper = vreinterpretq_u16_u8(in);
+    // 00000000000aaaaa
+    uint16x8_t upper_masked = vandq_u16(upper, vmovq_n_u16(0x1F));
+    // (effectively in >> 8)
+    // 110aaaaa10bbbbbb
+    uint16x8_t lower = vreinterpretq_u16_u8(vrev16q_u8(in));
+    // Assemble with shift left insert.
+    // 00000aaaaabbbbbb
+    uint16x8_t composed = vsliq_n_u16(lower, upper_masked, 6);
+    if (!match_system(big_endian)) composed = vreinterpretq_u16_u8(vrev16q_u8(vreinterpretq_u8_u16(composed)));
+    vst1q_u16(reinterpret_cast<uint16_t *>(utf16_output), composed);
     utf16_output += 8; // We wrote 16 bytes, 8 code points.
     return 16;
   }
   if(input_utf8_end_of_code_point_mask == 0x924) {
     // We want to take 4 3-byte UTF-8 words and turn them into 4 2-byte UTF-16 words.
-    // There is probably a more efficient sequence, but the following might do.
+    // This means that a more optimized table can be used.
+
+    // Low half contains  10cccccc|1110aaaa
+    // High half contains 10bbbbbb|10bbbbbb
 #ifdef SIMDUTF_REGULAR_VISUAL_STUDIO
-    const uint8x16_t sh = make_uint8x16_t(2, 1, 0, 255, 5, 4, 3, 255, 8, 7, 6, 255, 11, 10, 9, 255);
+    const uint8x16_t sh = make_uint8x16_t(0, 2, 3, 5, 6, 8, 9, 11, 1, 1, 4, 4, 7, 7, 10, 10);
 #else
-    const uint8x16_t sh = {2, 1, 0, 255, 5, 4, 3, 255, 8, 7, 6, 255, 11, 10, 9, 255};
+    const uint8x16_t sh = {0, 2, 3, 5, 6, 8, 9, 11, 1, 1, 4, 4, 7, 7, 10, 10};
 #endif
     uint8x16_t perm = vqtbl1q_u8(in, sh);
-    uint8x16_t ascii =
-        vandq_u8(perm, vreinterpretq_u8_u32(vmovq_n_u32(0x7f))); // 7 or 6 bits
-    uint8x16_t middlebyte =
-        vandq_u8(perm, vreinterpretq_u8_u32(vmovq_n_u32(0x3f00))); // 5 or 6 bits
-    uint8x16_t middlebyte_shifted = vreinterpretq_u8_u32(vshrq_n_u32(vreinterpretq_u32_u8(middlebyte), 2));
-    uint32x4_t highbyte =
-        vreinterpretq_u32_u8(vandq_u8(perm, vreinterpretq_u8_u32(vmovq_n_u32(0x0f0000)))); // 4 bits
-    uint32x4_t highbyte_shifted = vshrq_n_u32(highbyte, 4);
-    uint32x4_t composed =
-        vorrq_u32(vorrq_u32(vreinterpretq_u32_u8(ascii), vreinterpretq_u32_u8(middlebyte_shifted)), highbyte_shifted);
-    uint16x8_t composed_repacked = vmovn_high_u32(vmovn_u32(composed), composed);
-    if (!match_system(big_endian)) composed_repacked = vreinterpretq_u16_u8(vqtbl1q_u8(vreinterpretq_u8_u16(composed_repacked), swap));
-    vst1q_u16(reinterpret_cast<uint16_t*>(utf16_output), composed_repacked);
+    // Split
+    // 10cccccc|1110aaaa
+    uint8x8_t perm_low = vget_low_u8(perm); // no-op
+    // 10bbbbbb|10bbbbbb
+    uint8x8_t perm_high = vget_high_u8(perm);
+    // Equivalent to perm_low >> 8
+    // xxxxxxxx 10cccccc
+    uint16x4_t low = vreinterpret_u16_u8(vrev16_u8(perm_low));
+    // xxxxxxxx 10bbbbbb
+    uint16x4_t mid = vreinterpret_u16_u8(perm_high); // no-op
+    // xxxxxxxx 1110aaaa
+    uint16x4_t high = vreinterpret_u16_u8(perm_low); // no-op
+    // Assemble with shift left insert.
+    // xxxxxxaa aabbbbbb
+    uint16x4_t mid_high = vsli_n_u16(mid, high, 6);
+    // aaaabbbb bbcccccc
+    uint16x4_t composed = vsli_n_u16(low, mid_high, 6);
+    if (!match_system(big_endian)) composed= vreinterpret_u16_u8(vrev16_u8(vreinterpret_u8_u16(composed)));
+    vst1_u16(reinterpret_cast<uint16_t*>(utf16_output), composed);
     utf16_output += 4;
     return 12;
+  }
+  if(input_utf8_end_of_code_point_mask == 0x888) {
+    // We want to take 3-4 4-byte UTF-8 words and turn them into 3-4 4-byte UTF-16 pairs.
+    // Generating surrogate pairs is a little tricky though, but it is easier when we
+    // can assume they are all pairs.
+    size_t count = 3 + ((utf8_end_of_code_point_mask & 0xFFFF) == 0x8888);
+    // Swap byte pairs
+    // 10dddddd 10cccccc|10bbbbbb 11110aaa
+    // 10cccccc 10dddddd|11110aaa 10bbbbbb
+    uint8x16_t swap1 = vrev16q_u8(in);
+    // Shift left 2 bits
+    // cccccc00 dddddd00 xxxxxxxx bbbbbb00
+    uint32x4_t shift = vreinterpretq_u32_u8(vshlq_n_u8(swap1, 2));
+    // Insert low 2 bits of trail surrogate to magic number for later
+
+    // UTF-8 4b prefix  = -0xF000|0000
+    // surrogate offset = -0x0040|0000 (0x10000 << 6)
+    // surrogate high   = +0xD800|0000
+    // surrogate low    = +0x0000|DC00
+    // -------------------------------
+    //                  = +0xE7C0|DC00
+    // 11011100 00000000 11100111 110000cc
+    uint16x8_t magic = vreinterpretq_u16_u32(VSRAQ_N_U32(vmovq_n_u32(0xDC00E7C0), shift, 30));
+    // Generate unadjusted trail surrogate minus lowest 2 bits
+    // xxxxxxxx xxxxxxxx|11110aaa bbbbbb00
+    uint32x4_t trail = VBSLQ_U32(vmovq_n_u32(0x0000FF00), swap1, shift);
+    // Generate lead surrogate
+    // xxxxcccc ccdddddd|xxxxxxxx xxxxxxxx
+    uint32x4_t lead = vsliq_n_u16(swap1, in, 6);
+    // Mask out lead
+    // 000000cc ccdddddd|xxxxxxxx xxxxxxxx
+    uint32x4_t masked = vbicq_u32(lead, vmovq_n_u32(0xFC000000));
+    // Blend pairs
+    // 000000cc ccdddddd|11110aaa bbbbbb00
+    uint16x8_t blend = vreinterpretq_u16_u32(VBSLQ_U32(vmovq_n_u32(0x0000FFFF), trail, masked));
+    // Add magic number to correct UTF-8 prefix, surrogate offset, and surrogate prefixes.
+    // 110111CC CCDDDDDD|110110AA BBBBBBCC
+    uint16x8_t composed = vaddq_u16(blend, magic);
+    if (!match_system(big_endian)) composed = vreinterpretq_u16_u8(vrev16q_u8(vreinterpretq_u8_u16(composed)));
+    vst1q_u16(reinterpret_cast<uint16_t *>(utf16_output), composed);
+    utf16_output += count * 2;
+    return count * 4;
   }
   /// We do not have a fast path available, so we fallback.
 
@@ -88,92 +139,123 @@ size_t convert_masked_utf8_to_utf16(const char *input,
     // we process SIX (6) input code-words. The max length in bytes of six code
     // words spanning between 1 and 2 bytes each is 12 bytes.
     uint8x16_t sh = vld1q_u8(reinterpret_cast<const uint8_t*>(simdutf::tables::utf8_to_utf16::shufutf8[idx]));
-    uint8x16_t perm = vqtbl1q_u8(in, sh);
-    uint8x16_t ascii = vandq_u8(perm, vreinterpretq_u8_u16(vmovq_n_u16(0x7f)));
-    uint8x16_t highbyte = vandq_u8(perm, vreinterpretq_u8_u16(vmovq_n_u16(0x1f00)));
-    uint8x16_t composed = vorrq_u8(ascii, vreinterpretq_u8_u16(vshrq_n_u16(vreinterpretq_u16_u8(highbyte), 2)));
-    if (!match_system(big_endian)) composed = vqtbl1q_u8(composed, swap);
-    vst1q_u8(reinterpret_cast<uint8_t*>(utf16_output), composed);
+    uint16x8_t perm = vreinterpretq_u16_u8(vqtbl1q_u8(in, sh));
+    // Mask
+    // 1 byte: 00000000 0aaaaaaa
+    // 2 byte: 00000000 00bbbbbb
+    uint16x8_t ascii = vandq_u16(perm, vmovq_n_u16(0x7f)); // 6 or 7 bits
+    // 1 byte: 00000000 00000000
+    // 2 byte: 000aaaaa 00000000
+    uint16x8_t highbyte = vandq_u16(perm, vmovq_n_u16(0x1f00)); // 5 bits
+    // Combine with a shift right accumulate
+    // 1 byte: 00000000 0aaaaaaa
+    // 2 byte: 00000aaa aabbbbbb
+    uint16x8_t composed = vsraq_n_u16(ascii, highbyte, 2);
+    if (!match_system(big_endian)) composed = vreinterpretq_u16_u8(vrev16q_u8(vreinterpretq_u8_u16(composed)));
+    vst1q_u16(reinterpret_cast<uint16_t*>(utf16_output), composed);
     utf16_output += 6; // We wrote 12 bytes, 6 code points.
   } else if (idx < 145) {
     // FOUR (4) input code-words
     uint8x16_t sh = vld1q_u8(reinterpret_cast<const uint8_t*>(simdutf::tables::utf8_to_utf16::shufutf8[idx]));
-    uint8x16_t perm = vqtbl1q_u8(in, sh);
-    uint8x16_t ascii =
-        vandq_u8(perm, vreinterpretq_u8_u32(vmovq_n_u32(0x7f))); // 7 or 6 bits
-    uint8x16_t middlebyte =
-        vandq_u8(perm, vreinterpretq_u8_u32(vmovq_n_u32(0x3f00))); // 5 or 6 bits
-    uint8x16_t middlebyte_shifted = vreinterpretq_u8_u32(vshrq_n_u32(vreinterpretq_u32_u8(middlebyte), 2));
-    uint32x4_t highbyte =
-        vreinterpretq_u32_u8(vandq_u8(perm, vreinterpretq_u8_u32(vmovq_n_u32(0x0f0000)))); // 4 bits
-    uint32x4_t highbyte_shifted = vshrq_n_u32(highbyte, 4);
-    uint32x4_t composed =
-        vorrq_u32(vorrq_u32(vreinterpretq_u32_u8(ascii), vreinterpretq_u32_u8(middlebyte_shifted)), highbyte_shifted);
-    uint16x8_t composed_repacked = vmovn_high_u32(vmovn_u32(composed), composed);
-    if (!match_system(big_endian)) composed_repacked = vreinterpretq_u16_u8(vqtbl1q_u8(vreinterpretq_u8_u16(composed_repacked), swap));
-    vst1q_u16(reinterpret_cast<uint16_t*>(utf16_output), composed_repacked);
+    // XXX: depending on the system scalar instructions might be faster.
+    // 1 byte: 00000000 00000000 0ccccccc
+    // 2 byte: 00000000 110bbbbb 10cccccc
+    // 3 byte: 1110aaaa 10bbbbbb 10cccccc
+    uint32x4_t perm = vreinterpretq_u32_u8(vqtbl1q_u8(in, sh));
+    // Split into narrow vectors.
+    // This gives us a free shift for highperm.
+    // 3 byte:   aaaaxxxx xxxxxxxx
+    uint16x4_t highperm = vshrn_n_u32(perm, 4);
+    // 1 byte: 00000000 0ccccccc
+    // 2 byte: 110bbbbb 10cccccc
+    // 3 byte: 10bbbbbb 10cccccc
+    uint16x4_t lowperm = vmovn_u32(perm);
+    // ascii
+    uint16x4_t ascii = vand_u32(lowperm, vmov_n_u16(0x7F));
+    // 1 byte: 00000000 00000000
+    // 2 byte: 000bbbbb 00000000
+    // 3 byte: 000bbbbb 00000000
+    uint16x4_t middlebyte = vand_u16(lowperm, vmov_n_u16(0x3f00));
+    // Bitwise select the top and bottom bits
+    // 1 byte: 00000000 0ccccccc
+    // 2 byte: 00000000 00cccccc
+    // 3 byte: aaaa0000 00cccccc
+    uint16x4_t lowhigh = VBSL_U16(vmov_n_u16(0x0FFF), ascii, highperm);
+    // Shift right accumulate the middle byte
+    // 1 byte: 00000000 0ccccccc
+    // 2 byte: 00000bbb bbcccccc
+    // 3 byte: aaaabbbb bbcccccc
+    uint16x4_t composed = vsra_n_u16(lowhigh, middlebyte, 2);
+    if (!match_system(big_endian)) composed = vreinterpret_u16_u8(vrev16_u8(vreinterpret_u8_u16(composed)));
+    vst1_u16(reinterpret_cast<uint16_t*>(utf16_output), composed);
+
     utf16_output += 4;
   } else if (idx < 209) {
     // TWO (2) input code-words
-    //////////////
-    // There might be garbage inputs where a leading byte mascarades as a four-byte
-    // leading byte (by being followed by 3 continuation byte), but is not greater than
-    // 0xf0. This could trigger a buffer overflow if we only counted leading
-    // bytes of the form 0xf0 as generating surrogate pairs, without further UTF-8 validation.
-    // Thus we must be careful to ensure that only leading bytes at least as large as 0xf0 generate surrogate pairs.
-    // We do as at the cost of an extra mask.
-    /////////////
     uint8x16_t sh = vld1q_u8(reinterpret_cast<const uint8_t*>(simdutf::tables::utf8_to_utf16::shufutf8[idx]));
-    uint8x16_t perm = vqtbl1q_u8(in, sh);
-    uint8x16_t ascii = vandq_u8(perm, vreinterpretq_u8_u32(vmovq_n_u32(0x7f)));
-    uint8x16_t middlebyte = vandq_u8(perm, vreinterpretq_u8_u32(vmovq_n_u32(0x3f00)));
-    uint8x16_t middlebyte_shifted = vreinterpretq_u8_u32(vshrq_n_u32(vreinterpretq_u32_u8(middlebyte), 2));
-    uint8x16_t middlehighbyte = vandq_u8(perm, vreinterpretq_u8_u32(vmovq_n_u32(0x3f0000)));
-    // correct for spurious high bit
-    uint8x16_t correct =
-        vreinterpretq_u8_u32(vshrq_n_u32(vreinterpretq_u32_u8(vandq_u8(perm, vreinterpretq_u8_u32(vmovq_n_u32(0x400000)))), 1));
-    middlehighbyte = veorq_u8(correct, middlehighbyte);
-    uint8x16_t middlehighbyte_shifted = vreinterpretq_u8_u32(vshrq_n_u32(vreinterpretq_u32_u8(middlehighbyte), 4));
-    // We deliberately carry the leading four bits if they are present, we remove
-    // them later when computing hightenbits.
-    uint8x16_t highbyte = vandq_u8(perm, vreinterpretq_u8_u32(vmovq_n_u32(0xff000000)));
-    uint8x16_t highbyte_shifted = vreinterpretq_u8_u32(vshrq_n_u32(vreinterpretq_u32_u8(highbyte), 6));
-    // When we need to generate a surrogate pair (leading byte > 0xF0), then
-    // the corresponding 32-bit value in 'composed'  will be greater than
-    // > (0xff00000>>6) or > 0x3c00000. This can be used later to identify the
-    // location of the surrogate pairs.
-    uint8x16_t composed =
-        vorrq_u8(vorrq_u8(ascii, middlebyte_shifted),
-                     vorrq_u8(highbyte_shifted, middlehighbyte_shifted));
-    uint32x4_t composedminus =
-        vsubq_u32(vreinterpretq_u32_u8(composed), vmovq_n_u32(0x10000));
-    uint32x4_t lowtenbits =
-        vandq_u32(composedminus, vmovq_n_u32(0x3ff));
-    // Notice the 0x3ff mask:
-    uint32x4_t hightenbits = vandq_u32(vshrq_n_u32(composedminus, 10), vmovq_n_u32(0x3ff));
-    uint32x4_t lowtenbitsadd =
-        vaddq_u32(lowtenbits, vmovq_n_u32(0xDC00));
-    uint32x4_t hightenbitsadd =
-        vaddq_u32(hightenbits, vmovq_n_u32(0xD800));
-    uint32x4_t lowtenbitsaddshifted = vshlq_n_u32(lowtenbitsadd, 16);
+    // 1 byte: 00000000 00000000 00000000 0ddddddd
+    // 3 byte: 00000000 00000000 110ccccc 10dddddd
+    // 3 byte: 00000000 1110bbbb 10cccccc 10dddddd
+    // 4 byte: 11110aaa 10bbbbbb 10cccccc 10dddddd
+    uint32x4_t perm = vreinterpretq_u32_u8(vqtbl1q_u8(in, sh));
+    // If the high bit is 1 (s32 less than zero), this needs a surrogate pair
+    uint32x4_t is_pair = vreinterpretq_u32_s32(vcltzq_s32(vreinterpretq_s32_u32(perm)));
+
+    // Mask the low and middle bytes
+    // 00000000 00000000 00000000 0ddddddd
+    uint32x4_t ascii = vandq_u32(perm, vmovq_n_u32(0x7f));
+    // 00000000 00000000 00cccccc 00000000
+    uint32x4_t middlebyte = vandq_u32(perm, vmovq_n_u32(0x3F00));
+    // Because the surrogates need more work, the high surrogate is computed first.
+
+    // Start assembling the sequence. Since the 4th byte is in the same position as it
+    // would be in a surrogate and there is no dependency, shift left instead of right.
+    // 3 byte: 00000000 10bbbbxx xxxxxxxx xxxxxxxx
+    // 4 byte: 11110aaa bbbbbbxx xxxxxxxx xxxxxxxx
+    uint32x4_t ab = VBSLQ_U32(vmovq_n_u32(0xFF000000), perm, vshlq_n_u32(perm, 2));
+    // Top 16 bits contains the high ten bits of the surrogate pair before correction
+    // 3 byte: 00000000 10bbbbcc|cccc0000 00000000
+    // 4 byte: 11110aaa bbbbbbcc|cccc0000 00000000 - high 10 bits correct w/o correction
+    uint32x4_t abc = VBSLQ_U32(vmovq_n_u32(0xFFFC0000), ab, vshlq_n_u32(middlebyte, 4));
+    // Combine the low 6 or 7 bits by a shift right accumulate
+    // 3 byte: 00000000 00000010|bbbbcccc ccdddddd - low 16 bits correct
+    // 4 byte: 00000011 110aaabb|bbbbcccc ccdddddd - low 10 bits correct w/o correction
+    uint32x4_t composed = VSRAQ_N_U32(ascii, abc, 6);
+    // After this is for surrogates
+    // Blend the low and high surrogates
+    // 4 byte: 11110aaa bbbbbbcc|bbbbcccc ccdddddd
+    uint32x4_t mixed = VBSLQ_U32(vmovq_n_u32(0xFFFF0000), abc, composed);
+    // Clear the upper 6 bits of the low surrogate. Don't clear the upper bits yet as
+    // 0x10000 was not subtracted from the codepoint yet.
+    // 4 byte: 11110aaa bbbbbbcc|000000cc ccdddddd
+    uint16x8_t masked_pair = vreinterpretq_u16_u32(vbicq_u32(mixed, vmovq_n_u32(~0xFFFF03FF)));
+    // Correct the remaining UTF-8 prefix, surrogate offset, and add the surrogate prefixes
+    // in one magic addition.
+    // Number is halfword-swapped from above.
+    // 4 byte: 110110AA BBBBBBCC|110111CC CCDDDDDD - surrogate pair complete
     uint32x4_t surrogates =
-        vorrq_u32(hightenbitsadd, lowtenbitsaddshifted);
-    uint32_t basic_buffer[4];
-    uint32_t basic_buffer_swap[4];
+        vreinterpretq_u32_u16(vaddq_u16(masked_pair, vreinterpretq_u16_u32(vmovq_n_u32(0xE7C0DC00))));
+
+    // Select either the 4 byte surrogate pair or the 2 byte solo codepoint
+    // 3 byte: 0xxxxxxx xxxxxxxx|bbbbcccc ccdddddd
+    // 4 byte: 110110AA BBBBBBCC|110111CC CCDDDDDD
+    uint32x4_t selected = VBSLQ_U32(is_pair, surrogates, composed);
+    // Byte swap
     if (!match_system(big_endian)) {
-      vst1q_u32(basic_buffer_swap, vreinterpretq_u32_u8(vqtbl1q_u8(composed, swap)));
-      surrogates = vreinterpretq_u32_u8(vqtbl1q_u8(vreinterpretq_u8_u32(surrogates), swap));
+      selected = vreinterpretq_u32_u8(vrev16q_u8(vreinterpretq_u32_u8(selected)));
     }
-    vst1q_u32(basic_buffer, vreinterpretq_u32_u8(composed));
-    uint32_t surrogate_buffer[4];
-    vst1q_u32(surrogate_buffer, surrogates);
+    uint32_t buffer[4];
+    vst1q_u32(buffer, selected);
+    // Test for the top bit of the surrogate mask.
+    const uint32_t SURROGATE_MASK = match_system(big_endian) ? 0x80000000 : 0x00800000;
     for (size_t i = 0; i < 3; i++) {
-      if(basic_buffer[i] > 0x3c00000) {
-        utf16_output[0] = uint16_t(surrogate_buffer[i] & 0xffff);
-        utf16_output[1] = uint16_t(surrogate_buffer[i] >> 16);
+      // Surrogate
+      if (buffer[i] & SURROGATE_MASK) {
+        utf16_output[0] = uint16_t(buffer[i] >> 16);
+        utf16_output[1] = uint16_t(buffer[i] & 0xFFFF);
         utf16_output += 2;
       } else {
-        utf16_output[0] = !match_system(big_endian) ? uint16_t(basic_buffer_swap[i]) : uint16_t(basic_buffer[i]);
+        utf16_output[0] = uint16_t(buffer[i] & 0xFFFF);
         utf16_output++;
       }
     }
@@ -182,3 +264,4 @@ size_t convert_masked_utf8_to_utf16(const char *input,
   }
   return consumed;
 }
+