- Rewrite the intrinsics to be more NEON friendly
   - No longer a 1:1 translation of SSE4. While a generic algorithm is preferred, the common subset severely limits NEON's potential.
   - Use compound shifts, and bitwise selection to avoid `and; shr; orr`
      - Clang needs a hack to avoid constant propagating back to these.
   - Shuffle the 3 byte path in a way that is less clumsy for NEON (albeit a lot more confusing lol)
   - NEON has a very fast byte/word swap instruction, use it (applies to other functions too but I'm focusing here for now)
   - Make use of half vectors when appropriate (they can dual issue on little cores and have more favorable timing).
   - I can still do more micro-optimizations but I want to get feedback on this first. (I also want this before the half LUT, as I want the intrinsics to be optimized first)
 - Rework 4 byte sequence handling.
   - Completely rewritten.
   - Add a fast path for 3 or 4 4-byte sequences (the overhead for supporting both is like 2% since ARM doesn't need a branch). It is very worth it.
   - Calculate UTF-16 surrogates in parallel and skip the intermediate UTF-32
   - Use a blend mask to select the surrogate pairs
   - Use one magic addition to clear the UTF-8 prefix, adjust the surrogate offset, and add the surrogate prefixes
   - This new algorithm can be applied to SSE4/AVX2 with some tweaks.