@easyaspi314 2023-05-09
I do want to apply the 4 byte conversion rewrite to SSE4/AVX2 as well, the new approach is obviously faster from a rough draft, but I don't want to do too many things in one PR because my ADHD will go wild lol. (Also I am much less experienced in SSE anyways)

Do note that I plan on making more NEON optimization PRs.