@lemire 2023-05-09
> Do note that I plan on making more NEON optimization PRs.

They are certainly invited !!!

Of course, we want thorough benchmarks to make sure that it does move us forward.

I personally care very much about graviton 3 (neoverse) and Apple Silicon (+ Apple LLVM).