@lemire 2023-05-09
> While a generic algorithm is preferred

We are fine with dedicated algorithms. Of course, if there is no performance benefit to a specialized routine, we prefer to go generic... but when it pays to specialize, let us specialize.

I care a lot about ARM.