@easyaspi314 2023-05-09
I am still trying to work out some kinks in it. In theory it is better but GCC and Clang are both generating nonsense.

Issues:

 - Clang converts `usra` to `ushr; orr` or `ushr; add`
 - Clang converts `bsl` to `and; and; orr`
 - Inline assembly screws up instruction ordering
 - GCC reorders the compound in-place instructions in a way that forces it to use `mov`s everywhere.