@lemire 2023-05-09
I have avoided inline assembly, there is none in this project, deliberately.