@easyaspi314 2023-05-09
Noted. I might be able to get something passable with some simple empty asm statement guards, if that is okay. That can also be used to prevent GCC from reordering and screwing over the 4 byte paths. 

Also I was doing some tests. Apparently the slight slowdown is the branch for the 4 byte path. However, the dedicated path is **56% faster** when it comes to 3 or more 4 byte sequences. 

I can put it after the LUT load, which has some overhead from the redundant memory access but it is still much faster (43%) than not existing.