@lemire 2023-05-10
@easyaspi314 I recommend you do as you suggest, move it after the LUT. Then your PR becomes immediately an excellent candidate for merger... (I would need to read the code and do more benchmarks, more systems, more compilers, but I am quite optimistic that you already have something solid.) 

It always possible to come back for a second pass if you think it can be further improved.



Here is what I am getting on an M2 with the lipsum files... In three cases we get gains that are between ~5% and ~10%... which is fantastic if you ask me...

```
unicode_lipsum/lipsum/Arabic-Lipsum.utf8.txt
unicode_lipsum/lipsum/Chinese-Lipsum.utf8.txt
unicode_lipsum/lipsum/Emoji-Lipsum.utf8.txt
unicode_lipsum/lipsum/Hebrew-Lipsum.utf8.txt
unicode_lipsum/lipsum/Hindi-Lipsum.utf8.txt
unicode_lipsum/lipsum/Japanese-Lipsum.utf8.txt
unicode_lipsum/lipsum/Korean-Lipsum.utf8.txt
unicode_lipsum/lipsum/Latin-Lipsum.utf8.txt
unicode_lipsum/lipsum/Russian-Lipsum.utf8.txt
```

main:
```
   4.374 ins/byte,    1.659 cycle/byte,    2.178 GB/s (4.5 %),     3.614 GHz,    2.636 ins/cycle
   3.932 ins/byte,    0.743 cycle/byte,    5.079 GB/s (1.4 %),     3.776 GHz,    5.289 ins/cycle
   7.135 ins/byte,    1.154 cycle/byte,    3.084 GB/s (2.1 %),     3.560 GHz,    6.181 ins/cycle
   4.422 ins/byte,    1.667 cycle/byte,    2.177 GB/s (4.2 %),     3.630 GHz,    2.653 ins/cycle
   4.661 ins/byte,    1.535 cycle/byte,    2.347 GB/s (2.8 %),     3.601 GHz,    3.038 ins/cycle
   4.097 ins/byte,    0.866 cycle/byte,    4.317 GB/s (2.0 %),     3.740 GHz,    4.729 ins/cycle
   5.150 ins/byte,    1.707 cycle/byte,    2.126 GB/s (0.7 %),     3.628 GHz,    3.018 ins/cycle
   0.593 ins/byte,    0.109 cycle/byte,   50.902 GB/s (8.7 %),     5.560 GHz,    5.426 ins/cycle
   4.172 ins/byte,    1.614 cycle/byte,    2.221 GB/s (1.0 %),     3.586 GHz,    2.584 ins/cycle
```

PR:
```
   4.636 ins/byte,    1.701 cycle/byte,    2.126 GB/s (6.0 %),     3.618 GHz,    2.725 ins/cycle
   3.775 ins/byte,    0.719 cycle/byte,    5.271 GB/s (2.0 %),     3.792 GHz,    5.247 ins/cycle
   3.840 ins/byte,    1.064 cycle/byte,    3.480 GB/s (4.1 %),     3.701 GHz,    3.611 ins/cycle
   4.684 ins/byte,    1.710 cycle/byte,    2.122 GB/s (6.5 %),     3.629 GHz,    2.739 ins/cycle
   4.740 ins/byte,    1.566 cycle/byte,    2.303 GB/s (5.7 %),     3.608 GHz,    3.026 ins/cycle
   3.982 ins/byte,    0.816 cycle/byte,    4.610 GB/s (6.2 %),     3.760 GHz,    4.882 ins/cycle
   5.404 ins/byte,    1.719 cycle/byte,    2.100 GB/s (13.6 %),     3.611 GHz,    3.143 ins/cycle
   0.593 ins/byte,    0.131 cycle/byte,   39.375 GB/s (9.6 %),     5.174 GHz,    4.510 ins/cycle
   4.403 ins/byte,    1.646 cycle/byte,    2.177 GB/s (1.3 %),     3.583 GHz,    2.675 ins/cycle
```

Putting the new branch *after* the LUT, as suggested by @easyaspi314

```
   4.371 ins/byte,    1.659 cycle/byte,    2.173 GB/s (3.9 %),     3.606 GHz,    2.634 ins/cycle
   3.671 ins/byte,    0.701 cycle/byte,    5.407 GB/s (1.0 %),     3.793 GHz,    5.233 ins/cycle
   4.556 ins/byte,    1.109 cycle/byte,    3.375 GB/s (8.6 %),     3.742 GHz,    4.110 ins/cycle
   4.418 ins/byte,    1.668 cycle/byte,    2.177 GB/s (9.1 %),     3.631 GHz,    2.649 ins/cycle
   4.532 ins/byte,    1.493 cycle/byte,    2.419 GB/s (11.5 %),     3.612 GHz,    3.035 ins/cycle
   3.858 ins/byte,    0.808 cycle/byte,    4.650 GB/s (3.9 %),     3.755 GHz,    4.778 ins/cycle
   5.117 ins/byte,    1.673 cycle/byte,    2.160 GB/s (1.3 %),     3.613 GHz,    3.060 ins/cycle
   0.593 ins/byte,    0.123 cycle/byte,   43.470 GB/s (7.1 %),     5.351 GHz,    4.814 ins/cycle
   4.163 ins/byte,    1.612 cycle/byte,    2.227 GB/s (0.8 %),     3.591 GHz,    2.582 ins/cycle
```