@easyaspi314 2023-05-10
I found a few more adjustments that I am cleaning up now.

I didn't find much luck on Clang without using inline assembly, as frustrating as it is. Apparently the issue occured from two consecutive `vbsl` intrinsics that depend on each other though. 

But I agree, I can always optimize this later. There is a lot of low hanging fruit that I should aim for first. 😅