@lemire 2023-05-10
My benchmarking was done with clang, so your optimization works with clang without any asm. Maybe not as well as you'd like, but it works. Bumping up Chinese from 5 GB/s to 5.4 GB/s is fantastic.