@easyaspi314 2023-05-10
The asm is wrapped in those macros. 

I am truly baffled as to why the 3 byte path is consistently much faster than the 2 byte path though 🤔

Constants aside the pure two byte is
```
rev16
and
sli
```
compared to
```
tbl
ext
sli
rev16
sli
```
And it writes half the data.